requirements.txt
# Mistral AI SDK
mistralai>=1.0.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Structured output
pydantic>=2.0.0

# ElevenLabs TTS/STT
elevenlabs>=1.0.0

# assistant-transport protocol (optional)
assistant-stream>=0.0.1

# Redis (Upstash)
redis[hiredis]>=5.0.0

# Supabase
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.27.0
